This is the slowest test suite, runs for ~ 60s, because between every
document it adds 2048 "filler docs". This just adds up to a ton of
indexing across all the test methods.

Use 2048 for Nightly, and instead a smaller number (4) for local builds. It saves almost a minute of cpu time in tests.

Before:
```
The slowest suites (exceeding 1s) during this run:
  59.44s TestSimpleExplanationsWithFillerDocs (:lucene:core)
```

After: (no longer on the list < 9s for sure)
```
The slowest suites (exceeding 1s) during this run:
  14.98s TestSimpleTextDocValuesFormat (:lucene:codecs)
  14.06s TestLucene90DocValuesFormat (:lucene:core)
  13.96s TestLucene90DocValuesFormatMergeInstance (:lucene:core)
  13.64s TestFSTPostingsFormat (:lucene:codecs)
  11.41s TestPerFieldDocValuesFormat (:lucene:core)
  11.29s TestSimpleTextTermVectorsFormat (:lucene:codecs)
  11.10s TestAssertingDocValuesFormat (:lucene:test-framework)
  10.62s TestDirectPostingsFormat (:lucene:codecs)
   9.43s TestIndexSorting (:lucene:core)
   9.28s TestLatLonPointQueries (:lucene:core)
```